- Added protocol support for 15.21+ (GameLevelPercentU16, GameEffectSource, GameNpcWindowRedesign)
- New NPC trader module with HTML-based dialog and legacy fallback
- Client events system (popup banners for achievements, levels, skills, etc.)
- Per-source spell effect opacity (own, other player, creature, boss)
- Expanded cyclopedia offence stats with scrollable UI
- Game notifications module (info banner system)
- Fixed levelPercent U16 truncation bug (uint8_t -> uint16_t)
- Resource types aligned with Canary server
- Kept all local customizations (proficiency, task board, helper, etc.)
- No version gate on prey reroll price (Canary 15.23 always sends 4 bytes)

+++ src/client/localplayer.cpp
@@ -357,13 +357,13 @@ void LocalPlayer::setExperience(const uint64_t experience)
     callLuaField("onExperienceChange", experience, oldExperience);
 }
 
-void LocalPlayer::setLevel(const uint16_t level, const uint8_t levelPercent)
+void LocalPlayer::setLevel(const uint16_t level, const uint16_t levelPercent)
 {
     if (m_level == level && m_levelPercent == levelPercent)
         return;
 
     const uint16_t oldLevel = m_level;
-    const uint8_t oldLevelPercent = m_levelPercent;
+    const uint16_t oldLevelPercent = m_levelPercent;
 
     m_level = level;
     m_levelPercent = levelPercent;
