+++ src/client/localplayer.h
@@ -43,7 +43,7 @@ public:
     void setFreeCapacity(uint32_t freeCapacity);
     void setTotalCapacity(uint32_t totalCapacity);
     void setExperience(uint64_t experience);
-    void setLevel(uint16_t level, uint8_t levelPercent);
+    void setLevel(uint16_t level, uint16_t levelPercent);
     void setMana(uint32_t mana, uint32_t maxMana);
     void setManaShield(uint32_t manaShield, uint32_t maxManaShield);
     void setMagicLevel(uint16_t magicLevel, uint16_t magicLevelPercent);
@@ -80,7 +80,7 @@ public:
     uint16_t getMagicLevelPercent() { return m_magicLevelPercent; }
     uint16_t getBaseMagicLevel() { return m_baseMagicLevel; }
     uint8_t getSoul() { return m_soul; }
-    uint8_t getLevelPercent() { return m_levelPercent; }
+    uint16_t getLevelPercent() { return m_levelPercent; }
     uint8_t getHarmony() { return m_harmony; }
 
     uint16_t getLevel() { return m_level; }
@@ -208,7 +208,7 @@ private:
     uint32_t m_maxHealth{ 0 };
     uint64_t m_experience{ 0 };
     uint16_t m_level{ 0 };
-    uint8_t m_levelPercent{ 0 };
+    uint16_t m_levelPercent{ 0 };
     uint32_t m_mana{ 0 };
     uint32_t m_maxMana{ 0 };
     uint32_t m_manaShield{ 0 };
